Formatting changes, as usual.
MUProfile's method -openNewTelnetConnectionWithDelegate: was renamed to
-createNewTelnetConnectionWithDelegate:, and it no longer attempts to open the
connection in that factory method; that was assuming too much responsibility on
the part of the MUProfile. Specifically and in more detail:

The MUConnectionWindowController which has requested this telnet connection is
assigning it to a field in the object. The window controller is set as the
delegate for the telnet connection; therefore, during an attempted open the
delegate method will be called, which calls MUProfile's -loginWithConnection:
method with the value of that field. If the connection is opened by the factory
method on MUProfile, that field will not yet be assigned the value of the
created telnet connection, and autologin would be attempted with nil as the
value passed to -loginWithConnection: which would then naturally do nothing.

In MUConnectionWindowController's -connect: method, open the factory-created
J3NewTelnetConnection after it's correctly assigned the the telnetConnection
field.