Re #40 - changed J3Socket -readUpToLength: and -write: to handle EINTR correctly.
Also got really anal-retentive about using size_t and ssize_t as appropriate
when dealing with input to and results from read(2) and write(2) calls, and the
Obj-C methods that wrap them. This exposed a bug in the form of an incorrectly
handled return value in J3WriteBuffer -write, which manifests as a
signed/unsigned comparison.

This warning has been documented with a TODO and I'll file a new Trac ticket on
resolving all outstanding TODOs before 1.0.